Move SQL-related methods out from models
