+++ Cargo.toml
@@ -1,6 +1,5 @@
 [workspace]
 members = [
-  "prisma-models",
   "introspection-engine/core",
   "introspection-engine/connectors/introspection-connector",
   "introspection-engine/connectors/sql-introspection-connector",
@@ -17,6 +16,7 @@ members = [
   "libs/prisma-inflector",
   "libs/sql-schema-describer",
   "libs/sql-connection",
+  "libs/prisma-models",
 ]
 
 [patch.crates-io]
