+++ libs/datamodel/connectors/datamodel-connector/Cargo.toml
@@ -9,4 +9,3 @@ edition = "2018"
 [dependencies]
 chrono = { version = "0.4.6", features = ["serde"] }
 serde = { version = "1.0.90", features = ["derive"] }
-serde_json = "1.0"
