+++ libs/datamodel/core/Cargo.toml
@@ -5,7 +5,6 @@ authors = ["Emanuel Joebstl <emanuel.joebstl@gmail.com>"]
 edition = "2018"
 
 [dependencies]
-clap = "2.33.0"
 prisma-inflector = { path = "../../prisma-inflector" }
 datamodel-connector = { path = "../connectors/datamodel-connector" }
 # Temporary until PR is accepted.
@@ -18,7 +17,7 @@ serde_json = "1.0"
 failure = "0.1"
 failure_derive = "0.1"
 bytecount = "0.5"
-itertools = "0.8"
 
 [dev-dependencies]
 pretty_assertions = "0.6.1"
+clap = "2.33"
