+++ libs/datamodel/core/examples/dml-to-mcf.rs
@@ -1,7 +1,6 @@
 use datamodel;
 use std::fs;
 
-extern crate clap;
 use clap::{App, Arg};
 
 fn main() {
