+++ libs/prisma-models/src/field/relation.rs
@@ -1,7 +1,6 @@
 use super::FieldManifestation;
 use crate::prelude::*;
 use once_cell::sync::OnceCell;
-use quaint::ast::Column;
 use std::sync::{Arc, Weak};
 
 pub type RelationFieldRef = Arc<RelationField>;
@@ -120,6 +119,7 @@ impl RelationField {
             Some(RelationLinkManifestation::Inline(ref m)) => {
                 let is_self_rel = relation.is_self_relation();
 
+
                 if is_self_rel && self.is_hidden {
                     false
                 } else if is_self_rel && (self.relation_side == RelationSide::B || self.related_field().is_hidden) {
@@ -136,32 +136,6 @@ impl RelationField {
 
     pub fn relation_is_inlined_in_child(&self) -> bool {
         self.relation().is_inline_relation() && !self.relation_is_inlined_in_parent()
-    }
-
-    pub fn opposite_column(&self) -> Column<'static> {
-        match self.relation_side {
-            RelationSide::A => self.relation().model_b_column(),
-            RelationSide::B => self.relation().model_a_column(),
-        }
-    }
-
-    pub fn relation_column(&self) -> Column<'static> {
-        match self.relation_side {
-            RelationSide::A => self.relation().model_a_column(),
-            RelationSide::B => self.relation().model_b_column(),
-        }
-    }
-
-    pub fn as_column(&self) -> Column<'static> {
-        let model = self.model();
-        let internal_data_model = model.internal_data_model();
-        let db_name = self.db_name();
-        let parts = (
-            (internal_data_model.db_name.clone(), model.db_name().to_string()),
-            db_name.clone(),
-        );
-
-        parts.into()
     }
 
     pub fn related_model(&self) -> ModelRef {
