+++ libs/prisma-models/src/sql_ext.rs
@@ -1,0 +1,15 @@
+mod column;
+mod table;
+mod relation;
+mod selected_fields;
+mod scalar_list_table;
+mod scalar_field;
+mod prisma_value;
+
+pub use column::*;
+pub use table::*;
+pub use relation::*;
+pub use selected_fields::*;
+pub use scalar_list_table::*;
+pub use scalar_field::*;
+pub use prisma_value::*;
