+++ libs/sql-schema-describer/Cargo.toml
@@ -10,7 +10,6 @@ chrono = { version = "0.4" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 quaint = { git = "https://github.com/prisma/quaint.git" }
-itertools = "0.8"
 url = "1.7.2"
 log = "0.4"
 regex = "1.2"
