+++ migration-engine/connectors/sql-migration-connector/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2018"
 migration-connector = { path = "../migration-connector" }
 sql-schema-describer = { path = "../../../libs/sql-schema-describer" }
 datamodel = { path = "../../../libs/datamodel/core" }
-prisma-models = { path = "../../../prisma-models" }
+prisma-models = { path = "../../../libs/prisma-models" }
 chrono = { version = "0.4" }
 serde_json = "1.0"
 serde = "1.0"
