+++ query-engine/connectors/query-connector/Cargo.toml
@@ -8,11 +8,13 @@ edition = "2018"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 once_cell = "1.2"
-quaint = { git = "https://github.com/prisma/quaint.git" }
-prisma-models = { path = "../../../prisma-models" }
+prisma-models = { path = "../../../libs/prisma-models" }
 failure = "0.1"
 failure_derive = "0.1"
 uuid = "0.7"
 itertools = "0.8"
 chrono = { version = "0.4", features = ["serde"] }
 futures = "0.3"
+
+[dev-dependencies]
+serde_json = "1"
