+++ query-engine/connectors/query-connector/src/filter/relation.rs
@@ -37,7 +37,6 @@ impl RelationCompare for Arc<RelationField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
     /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
@@ -88,7 +87,6 @@ impl RelationCompare for Arc<RelationField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
     /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
@@ -139,7 +137,6 @@ impl RelationCompare for Arc<RelationField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
     /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
@@ -190,7 +187,6 @@ impl RelationCompare for Arc<RelationField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
     /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
@@ -241,7 +237,6 @@ impl RelationCompare for Arc<RelationField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
     /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
