+++ query-engine/connectors/query-connector/src/filter/scalar.rs
@@ -54,8 +54,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -87,8 +85,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -120,8 +116,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -153,8 +147,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -186,8 +178,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -219,8 +209,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -252,8 +240,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -285,8 +271,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -318,8 +302,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -351,8 +333,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -384,8 +364,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
@@ -417,8 +395,6 @@ impl ScalarCompare for Arc<ScalarField> {
     /// ```rust
     /// # use query_connector::{*, filter::*};
     /// # use prisma_models::*;
-    /// # use quaint::ast::*;
-    /// # use serde_json;
     /// # use std::{fs::File, sync::Arc};
     /// #
     /// # let tmp: InternalDataModelTemplate = serde_json::from_reader(File::open("../sql-query-connector/test_schema.json").unwrap()).unwrap();
